debt - use main in main side services consistently (workspaces)
