--- src/vs/platform/workspaces/electron-main/workspacesService.ts
@@ -1,81 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { AddFirstParameterToFunctions } from 'vs/base/common/types';
-import { IWorkspacesService, IEnterWorkspaceResult, IWorkspaceFolderCreationData, IWorkspaceIdentifier, IRecentlyOpened, IRecent } from 'vs/platform/workspaces/common/workspaces';
-import { URI } from 'vs/base/common/uri';
-import { IWorkspacesMainService } from 'vs/platform/workspaces/electron-main/workspacesMainService';
-import { IWindowsMainService } from 'vs/platform/windows/electron-main/windows';
-import { IWorkspacesHistoryMainService } from 'vs/platform/workspaces/electron-main/workspacesHistoryMainService';
-import { IBackupMainService } from 'vs/platform/backup/electron-main/backup';
-
-export class WorkspacesService implements AddFirstParameterToFunctions<IWorkspacesService, Promise<unknown> /* only methods, not events */, number /* window ID */> {
-
-	declare readonly _serviceBrand: undefined;
-
-	constructor(
-		@IWorkspacesMainService private readonly workspacesMainService: IWorkspacesMainService,
-		@IWindowsMainService private readonly windowsMainService: IWindowsMainService,
-		@IWorkspacesHistoryMainService private readonly workspacesHistoryMainService: IWorkspacesHistoryMainService,
-		@IBackupMainService private readonly backupMainService: IBackupMainService
-	) {
-	}
-
-	//#region Workspace Management
-
-	async enterWorkspace(windowId: number, path: URI): Promise<IEnterWorkspaceResult | null> {
-		const window = this.windowsMainService.getWindowById(windowId);
-		if (window) {
-			return this.workspacesMainService.enterWorkspace(window, this.windowsMainService.getWindows(), path);
-		}
-
-		return null;
-	}
-
-	createUntitledWorkspace(windowId: number, folders?: IWorkspaceFolderCreationData[], remoteAuthority?: string): Promise<IWorkspaceIdentifier> {
-		return this.workspacesMainService.createUntitledWorkspace(folders, remoteAuthority);
-	}
-
-	deleteUntitledWorkspace(windowId: number, workspace: IWorkspaceIdentifier): Promise<void> {
-		return this.workspacesMainService.deleteUntitledWorkspace(workspace);
-	}
-
-	getWorkspaceIdentifier(windowId: number, workspacePath: URI): Promise<IWorkspaceIdentifier> {
-		return this.workspacesMainService.getWorkspaceIdentifier(workspacePath);
-	}
-
-	//#endregion
-
-	//#region Workspaces History
-
-	readonly onRecentlyOpenedChange = this.workspacesHistoryMainService.onRecentlyOpenedChange;
-
-	async getRecentlyOpened(windowId: number): Promise<IRecentlyOpened> {
-		return this.workspacesHistoryMainService.getRecentlyOpened(this.windowsMainService.getWindowById(windowId));
-	}
-
-	async addRecentlyOpened(windowId: number, recents: IRecent[]): Promise<void> {
-		return this.workspacesHistoryMainService.addRecentlyOpened(recents);
-	}
-
-	async removeRecentlyOpened(windowId: number, paths: URI[]): Promise<void> {
-		return this.workspacesHistoryMainService.removeRecentlyOpened(paths);
-	}
-
-	async clearRecentlyOpened(windowId: number): Promise<void> {
-		return this.workspacesHistoryMainService.clearRecentlyOpened();
-	}
-
-	//#endregion
-
-
-	//#region Dirty Workspaces
-
-	async getDirtyWorkspaces(): Promise<Array<IWorkspaceIdentifier | URI>> {
-		return this.backupMainService.getDirtyWorkspaces();
-	}
-
-	//#endregion
-}
